数据存储从coreData改用SQLite.swift

+++ MyClear/myListCell_textField.swift
@@ -9,6 +9,6 @@ import UIKit
 
 class myListCell_textField: UITextField {
     
-    var id : Int32?
+    var id : Int64?
 
 }
